+++ native/src/main/java/io/ballerina/stdlib/ftp/server/FtpListenerHelper.java
@@ -682,12 +682,16 @@ public class FtpListenerHelper {
         }
     }
 
-    private static void configureServerFtpsSecureSocket(BMap secureSocket, Map<String, Object> params) 
+    private static void configureServerFtpsSecureSocket(BMap secureSocket, Map<String, Object> params)
             throws BallerinaFtpException {
         FtpUtil.configureFtpsMode(secureSocket, params);
         FtpUtil.configureFtpsDataChannelProtection(secureSocket, params);
-        
-        String keyStorePath = FtpUtil.extractAndConfigureStore(secureSocket, FtpConstants.SECURE_SOCKET_KEY, 
+
+        boolean verifyHostName = secureSocket.getBooleanValue(
+                StringUtils.fromString(FtpConstants.ENDPOINT_CONFIG_VERIFY_HOST_NAME));
+        params.put(FtpConstants.ENDPOINT_CONFIG_VERIFY_HOST_NAME, verifyHostName);
+
+        String keyStorePath = FtpUtil.extractAndConfigureStore(secureSocket, FtpConstants.SECURE_SOCKET_KEY,
                 FtpConstants.ENDPOINT_CONFIG_KEYSTORE_PATH, 
                 FtpConstants.ENDPOINT_CONFIG_KEYSTORE_PASSWORD, 
                 params);